Add files via upload
Data Visualization Example of the daily tweets of both houses of Congress (plus joint committees), encompassing 1,000+ campaign, office, committee and party accounts.
The provided data is for 11-05-2018 and 11-04-2018. 
The Gephi file is based on who mentioned whom example. 
The next version will contain who mentioned whom with which sentiment.
Be ready :)